requirements.txt
# Core dependencies
torch>=1.9.0
numpy>=1.21.0
PyYAML>=5.4.1
tqdm>=4.62.0

# Visualization dependencies
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine learning dependencies
scikit-learn>=1.0.0
scipy>=1.7.0

# Logging and monitoring
wandb>=0.12.0

# Optional dependencies for extended functionality
librosa>=0.9.0  # For advanced audio processing
jiwer>=2.3.0    # For word error rate computation
nltk>=3.6       # For text processing

# Development dependencies
pytest>=6.2.0
black>=21.0.0
flake8>=4.0.0